Separates fluctuation to its own module
All related things to the fluctuation feature will be handle in this module from now on.
